pyproject.toml
[project]
name = "avni-ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "openai>=1.99.9",
    "fastmcp>=2.11.3",
    "fastapi>=0.118.0",
    "pytest-asyncio>=1.1.0",
    "dspy-ai>=3.0.3",
    "requests>=2.32.4",
    "pandas>=2.3.3",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "openpyxl>=3.1.5",
    "pymongo>=4.16.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
avni-ai-server = "src.main:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-ordering>=0.6",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

dev = [
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "asyncio>=4.0.0",
    "anyio>=4.9.0",
]
